[Chore #163461416] Delete all unused or duplicate files and folders from the repository branch
