+++ UI/css/main.css
@@ -2,17 +2,22 @@
 
 html, body, .container {
   width: 100%;
-  height: 100%;
+}
+
+.container, .container-flex {
+  display: flex;
+  flex-direction: column;
 }
 
 body {
   padding: 0;
   margin: 0;
-  background-color: #5d5d5d;
+  background-color: #f5f5f5;
+  min-height: 100vh;
 }
 
 .nav {
-  background-color: #efefef;
+  background-color: #c20000;
   height: 50px;
   box-sizing: border-box;
   box-shadow: 0 5px 10px #000;
@@ -46,7 +51,7 @@ body {
 }
 
 .nav > ul > li > a {
-  color: #222;
+  color: #fffbfb;
   font-family: Montserrat, sans-serif;
   text-decoration: none;
   height: 50px;
@@ -58,7 +63,11 @@ body {
 }
 
 .nav > ul > li > a:hover {
-  background-color: #dfdfdf;
+  background-color: #646363;
+}
+
+.nav > ul > li > a:active, .nav > ul > li > a.active {
+  background-color: #3f3f3f;
 }
 
 .nav.active > ul > li > a:hover {
@@ -70,7 +79,7 @@ body {
 
 .bg {
   width: 100%;
-  height: 100vh;
+  min-height: 100vh;
   background-attachment: fixed;
   background-position: center;
   background-repeat: no-repeat;
@@ -177,6 +186,9 @@ input,select {
 }
 
 input[type='text'],
+textarea,
+input[type='textarea'],
+input[type='textbox'],
 input[type='password'],
 input[type='number'],
 input[type='date'],
@@ -184,7 +196,7 @@ input[type='file'],
 select{
     display: block;
     width: 100%;
-    height: 2rem;
+    min-height: 2rem;
     padding: 0 0.75rem;
     border: 1px solid hsl(0, 0%, 75%);
     border: 1px solid black;
@@ -192,18 +204,20 @@ select{
     color: black;
 }
 
+button {
+  border-style: none;
+}
+
 .button-default{
-  border-radius: 2px;
   padding: 12px;
   margin: 4px;
 }
 
-.view-quest,.myButton,.button-default,.edit-quest,.delete-quest {
+.update-status-quest,.view-quest,.myButton,.button-default,.edit-quest,.delete-quest {
 color: white;
 font-size: 16px;
 line-height: 10px;
 padding: 21px;
-border-radius: 10px;
 font-weight: normal;
 text-decoration: none;
 font-style: normal;
@@ -213,29 +227,35 @@ background: #020202;
 }
 
 .view-quest:hover,.myButton:hover,.button-default:hover {
-background: #1a1a1a; }
+background: #706f6f;
+}
+
 .view-quest:active,.myButton:active,.button-default:active {
-background: #252525; }
+  background: #484848;
+  border-style: solid;
+  border-width: 1px;
+  border-color: #7a7a7a;
+}
 
 .view-quest{
   padding: 11px;
   margin:2px;
 }
 
-.edit-quest {
+.update-status-quest,.edit-quest {
   background-color: #0084c3;
   color: white;
   padding: 11px;
   margin:2px;
 }
 
-.edit-quest:hover {
+.edit-quest:hover,.update-status-quest:hover {
   background-color: #003e5c;
   color: white;
 
 }
 
-.edit-quest:active {
+.edit-quest:active, .update-status-quest:active {
   background-color: #003e5c;
   color: white;
 
@@ -273,9 +293,9 @@ table {
 }
 
 td, th {
-    border: 1px solid #dddddd;
-    text-align: left;
-    padding: 8px;
+  border: 1px solid #dddddd;
+  text-align: left;
+  padding: 14px;
 }
 
 .button-default>a,.edit-quest>a,.delete-quest>a,.view-quest>a{
@@ -287,13 +307,87 @@ td, th {
   width: 90%;
 }
 
-.map-container{
-  height:400px;
-  width:550px;
+/* Google maps api css */
+#description {
+  font-family: Roboto;
+  font-size: 15px;
+  font-weight: 300;
+}
+
+#infowindow-content .title {
+  font-weight: bold;
+}
+
+#infowindow-content {
+  display: none;
+}
+
+#map #infowindow-content {
+  display: inline;
+}
+
+.pac-card {
+  margin: 10px 10px 0 0;
+  border-radius: 2px 0 0 2px;
+  box-sizing: border-box;
+  -moz-box-sizing: border-box;
+  outline: none;
+  box-shadow: 0 2px 6px rgba(0, 0, 0, 0.3);
+  background-color: #fff;
+  font-family: Roboto;
+}
+
+#pac-container {
+  padding-bottom: 12px;
+  margin-right: 12px;
+}
+
+.pac-controls {
+  display: inline-block;
+  padding: 5px 11px;
+}
+
+.pac-controls label {
+  font-family: Roboto;
+  font-size: 13px;
+  font-weight: 300;
+}
+
+#pac-input {
+  background-color: #fff;
+  font-family: Roboto;
+  font-size: 15px;
+  font-weight: 300;
+  margin-left: 12px;
+  padding: 0 11px 0 13px;
+  text-overflow: ellipsis;
+  width: 400px;
+}
+
+#pac-input:focus {
+  border-color: #4d90fe;
+}
+
+#title {
+  color: #fff;
+  background-color: #4d90fe;
+  font-size: 25px;
+  font-weight: 500;
+  padding: 6px 12px;
+}
+#target {
+  width: 345px;
+}
+
+.map-container {
+    height: 400px;
+    width: 500px;
 }
 #map {
   height: 100%;
 }
+
+/* Google maps api css end */
 
 #alert_error,#alert_success,#alert_loading {
   position: fixed;
@@ -320,4 +414,17 @@ td, th {
 }
 #alert_loading {
   color: black;
+}
+
+.container-flex{
+  width: 90%;
+  margin-left: 5%;
+  margin-right: 5%;
+  min-height: 100%;
+  justify-content: center;
+  align-items: center;
+}
+
+td, th {
+  max-width: 250px;
 }
