+++ UI/js/get_all_incidents.js
