+++ UI/js/view_incident.js
