+++ UI/profile.html
@@ -13,11 +13,11 @@
         <li><a href="index.html">Home</a></li>
         <li><a href="view_my_quests.html">My Red Flags/Interventions</a></li>
         <li><a href="create_quest.html">Create an incident</a></li>
-        <li class="li-right"><a id="logout_link" onclick="logout()">Logout</a></li>
+        <li class="li-right"><a id="logout_link" >Logout</a></li>
       </ul>
 </div>
 <div class="container">
-  <div id="profile" class="bg">
+  <div id="profile" class="container-flex">
       <div class="form-container">
             <h1>Your details</h1>
               <div class="update_user_form" >
@@ -56,5 +56,6 @@
     <li><a href="#login">Login</a></li>
   </div>
 </div>
+<script src="https://maps.googleapis.com/maps/api/js?key=AIzaSyDLjJJ8MEg85s3CU2eyQdZ_4H5Uaz0oud0&libraries=places"></script>
   </body>
 </html>
