+++ UI/view_my_quests.html
@@ -6,6 +6,7 @@
     <meta charset="utf-8">
     <title>iReporter</title>
     <link rel="stylesheet" href="./css/main.css">
+    <script type="module" src="js/get_all_incidents.js"></script>
     <script type="module" src="js/main.js"></script>
   </head>
   <body>
@@ -14,7 +15,7 @@
         <li><a href="index.html">Home</a></li>
         <li><a href="profile.html">My Profile</a></li>
         <li><a href="create_quest.html">Create an incident</a></li>
-        <li class="li-right"><a id="logout" href="">Logout</a></li>
+        <li class="li-right"><a id="logout_link" href="">Logout</a></li>
       </ul>
 </div>
 <div class="container">
@@ -26,10 +27,10 @@
       <table>
         <thead>
             <tr>
-                <th>Resolved Red-flags/Interventions </th>
-                <th>Rejected Red-flags/Interventions</th>
-                <th>Under Investigation Red-flags/Interventions</th>
-                <th>Draft Red-flad/Interventions</th>
+                <th>Resolved Incidents</th>
+                <th>Rejected Incidents</th>
+                <th>Under Incidents</th>
+                <th>Draft Incidents</th>
             </tr>
         </thead>
         <tbody>
@@ -37,25 +38,26 @@
                 <td>0</td>
                 <td>0</td>
                 <td>0</td>
-                <td>3</td>
+                <td>0</td>
             </tr>
         <tbody>
     </table>
     </div>
     <h1>My Red-flags & Interventions</h1>
     <div class="">
-      <b>Create a red-flag/intervention</b>
-      <button class="button-default"><a href="#" > Create </a></button>
-    <br>
+      <h2><b>Create an Incident</b></h2>
+      <a href="create_quest.html" class="button-default" > Create </a>
     </div>
+    <br>
     <div>
       <div class="form-group">
-        <label for="Status Filter"><b>Red-flag/Intervention Status Filter : </b> </label>
+        <label for="Status Filter"><b>Incident Status Filter : </b> </label>
         <select>
-          <option value="1">Draft</option>
-          <option value="2">Resolved</option>
-          <option value="3">Rejected</option>
-          <option value="4">Under Investigation</option>
+          <option value="all">All</option>
+          <option value="draft">Draft</option>
+          <option value="resolved">Resolved</option>
+          <option value="rejected">Rejected</option>
+          <option value="under investigation">Under Investigation</option>
         </select>
       </div>
     </div>
@@ -63,40 +65,24 @@
     <table>
         <thead>
             <tr>
-                <th>Red-flag/Intervetion Name</th>
+                <th>Incident Title</th>
                 <th>Type</th>
                 <th>Status</th>
                 <th>Actions</th>
             </tr>
         </thead>
-        <tbody>
-            <tr class=" ">
-                <td>Bribery in Procurement</td>
-                <td>Red-flag</td>
-                <td>Draft</td>
-                <td><button class="edit-quest"><a href="">Edit</a></button><button class="delete-quest"><a href="">Delete</a></button></td>
-            </tr>
-            <tr class=" ">
-                <td>Pothole or highway</td>
-                <td>Intervention</td>
-                <td>Draft</td>
-                <td><button class="edit-quest"><a href="">Edit</a></button><button class="delete-quest"><a href="">Delete</a></button></td>
-            </tr>
-            <tr class=" ">
-                <td>Electric pole on road</td>
-                <td>Intervention</td>
-                <td>Under Investigation</td>
-                <td><button class="view-quest"><a href="">View</a></button></td>
-            </tr>
+        <tbody id="incidents_tbody">
         <tbody>
     </table>
   </div>
 </div>
   <div class="footer">
-    <li><a href="#home">Home</a></li>
-    <li><a href="#quests">My Red Flags/Interventions</a></li>
-    <li class="li-right"><a href="#logout">Logout</a></li>
+    <li><a href="index.html">Home</a></li>
+    <li><a href="profile.html">My Profile</a></li>
+    <li><a href="create_quest.html">Create an incident</a></li>
+    <li class="li-right"><a id="logout_link" href="">Logout</a></li>
   </div>
 </div>
+<script src="https://maps.googleapis.com/maps/api/js?key=AIzaSyDLjJJ8MEg85s3CU2eyQdZ_4H5Uaz0oud0&libraries=places"></script>
   </body>
 </html>
