--- css/main.css
@@ -1,430 +1,0 @@
-@import url("https://fonts.googleapis.com/css?family=Montserrat:400,400i,700");
-
-html, body, .container {
-  width: 100%;
-}
-
-.container, .container-flex {
-  display: flex;
-  flex-direction: column;
-}
-
-body {
-  padding: 0;
-  margin: 0;
-  background-color: #f5f5f5;
-  min-height: 100vh;
-}
-
-.nav {
-  background-color: #c20000;
-  height: 50px;
-  box-sizing: border-box;
-  box-shadow: 0 5px 10px #000;
-  position: relative;
-  top: 0;
-  left: 0;
-  right: 0;
-  transition: top 0.15s ease-out, border-radius 0.15s ease-out;
-}
-
-.nav.active {
-  background-color: rgba(255, 255, 255, 0.8);
-  position: fixed;
-  top: 20px;
-  left: 20px;
-  right: 20px;
-  border-radius: 8px 5px;
-}
-
-.nav > ul {
-  width: 100%;
-  padding: 0 20px;
-  margin: 0;
-  box-sizing: border-box;
-  list-style-type: none;
-}
-
-.nav > ul > li {
-  float: left;
-  display: inline-block;
-}
-
-.nav > ul > li > a {
-  color: #fffbfb;
-  font-family: Montserrat, sans-serif;
-  text-decoration: none;
-  height: 50px;
-  line-height: 50px;
-  padding: 0 20px;
-  display: block;
-  text-shadow: 0 -1px #fff;
-  transition: background 0.15s ease;
-}
-
-.nav > ul > li > a:hover {
-  background-color: #646363;
-}
-
-.nav > ul > li > a:active, .nav > ul > li > a.active {
-  background-color: #3f3f3f;
-}
-
-.nav.active > ul > li > a:hover {
-  background-color: rgba(180, 180, 180, 0.7);
-}
-.li-right{
-  float: right !important;
-}
-
-.bg {
-  width: 100%;
-  min-height: 100vh;
-  background-attachment: fixed;
-  background-position: center;
-  background-repeat: no-repeat;
-  background-size: cover;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-.bg-alt {
-  width: 100%;
-  background-attachment: fixed;
-  background-position: center;
-  background-repeat: no-repeat;
-  background-size: cover;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-.bg> div ,.bg-alt>div{
-    color: white;
-    font-family: Montserrat, sans-serif;
-    text-align: center;
-    width: 100%;
-    padding: 30px;
-    text-shadow: 0 -1px #222;
-}
-
-#home {
-  background-image: linear-gradient(rgba(0, 0, 0, 0.6), rgba(0, 0, 0, 0.6)), url('../images/ivmpunity.png');
-}
-
-#register {
-  background-image: linear-gradient(rgba(0, 0, 0, 0.6), rgba(0, 0, 0, 0.6)), url('../images/blackunity.jpg');
-}
-
-#login {
-  background-image: linear-gradient(rgba(0, 0, 0, 0.6), rgba(0, 0, 0, 0.6)), url('../images/stopcorruption.jpg');
-}
-
-
-.footer {
-  font-family: Montserrat, sans-serif;
-  background-color: #efefef;
-  width: 100%;
-  height: 200px;
-  padding: 30px;
-  display: flex;
-  align-items: flex-end;
-  align-content: flex-end;
-  flex-wrap: wrap;
-  box-sizing: border-box;
-}
-
-.footer > span:first-child {
-  font-weight: bold;
-  margin-bottom: 25px;
-}
-
-.footer > span:last-child {
-  margin-top: 25px;
-}
-
-.footer > span {
-  color: #222;
-  width: 100%;
-  display: block;
-  margin-top: 10px;
-}
-
-.footer > span > a {
-  color: #888;
-  text-decoration: none;
-}
-
-.footer > span > a:hover {
-  text-decoration: underline;
-}
-
-.form-group {
-    position: relative;
-    padding-top: 1.5rem;
-}
-
-.form-container{
-  max-width: 500px;
-}
-
-#quest_table{
-  background-color: #141414;
-}
-
-label {
-    opacity: 1;
-    transform: translateY(0);
-    transition: all 0.2s ease-out;
-}
-
-input,select {
-    background-repeat: no-repeat;
-    background-position: right 1rem center;
-    background-size: 0.75rem;
-}
-
-input[type='text'],
-textarea,
-input[type='textarea'],
-input[type='textbox'],
-input[type='password'],
-input[type='number'],
-input[type='date'],
-input[type='file'],
-select{
-    display: block;
-    width: 100%;
-    min-height: 2rem;
-    padding: 0 0.75rem;
-    border: 1px solid hsl(0, 0%, 75%);
-    border: 1px solid black;
-    color: hsl(0, 0%, 25%);
-    color: black;
-}
-
-button {
-  border-style: none;
-}
-
-.button-default{
-  padding: 12px;
-  margin: 4px;
-}
-
-.update-status-quest,.view-quest,.myButton,.button-default,.edit-quest,.delete-quest {
-color: white;
-font-size: 16px;
-line-height: 10px;
-padding: 21px;
-font-weight: normal;
-text-decoration: none;
-font-style: normal;
-font-variant: normal;
-text-transform: none;
-background: #020202;
-}
-
-.view-quest:hover,.myButton:hover,.button-default:hover {
-background: #706f6f;
-}
-
-.view-quest:active,.myButton:active,.button-default:active {
-  background: #484848;
-  border-style: solid;
-  border-width: 1px;
-  border-color: #7a7a7a;
-}
-
-.view-quest{
-  padding: 11px;
-  margin:2px;
-}
-
-.update-status-quest,.edit-quest {
-  background-color: #0084c3;
-  color: white;
-  padding: 11px;
-  margin:2px;
-}
-
-.edit-quest:hover,.update-status-quest:hover {
-  background-color: #003e5c;
-  color: white;
-
-}
-
-.edit-quest:active, .update-status-quest:active {
-  background-color: #003e5c;
-  color: white;
-
-}
-
-.delete-quest {
-  background-color: #c50000;
-  color: white;
-  padding: 11px;
-  margin:2px;
-}
-
-.delete-quest:hover {
-  background-color: #7b0000;
-  color: white;
-}
-
-.delete-quest:active {
-  background-color: #7b0000;
-  color: white;
-}
-
-input:placeholder-shown + label {
-    opacity: 0;
-    transform: translateY(1rem);
-}
-
-#quest_create,#quest_edit{
-      background-color: #250303;
-}
-
-table {
-    border-collapse: collapse;
-    width: 100%;
-}
-
-td, th {
-  border: 1px solid #dddddd;
-  text-align: left;
-  padding: 14px;
-}
-
-.button-default>a,.edit-quest>a,.delete-quest>a,.view-quest>a{
-  text-decoration: none;
-  color: inherit;
-}
-
-.table-container {
-  width: 90%;
-}
-
-/* Google maps api css */
-#description {
-  font-family: Roboto;
-  font-size: 15px;
-  font-weight: 300;
-}
-
-#infowindow-content .title {
-  font-weight: bold;
-}
-
-#infowindow-content {
-  display: none;
-}
-
-#map #infowindow-content {
-  display: inline;
-}
-
-.pac-card {
-  margin: 10px 10px 0 0;
-  border-radius: 2px 0 0 2px;
-  box-sizing: border-box;
-  -moz-box-sizing: border-box;
-  outline: none;
-  box-shadow: 0 2px 6px rgba(0, 0, 0, 0.3);
-  background-color: #fff;
-  font-family: Roboto;
-}
-
-#pac-container {
-  padding-bottom: 12px;
-  margin-right: 12px;
-}
-
-.pac-controls {
-  display: inline-block;
-  padding: 5px 11px;
-}
-
-.pac-controls label {
-  font-family: Roboto;
-  font-size: 13px;
-  font-weight: 300;
-}
-
-#pac-input {
-  background-color: #fff;
-  font-family: Roboto;
-  font-size: 15px;
-  font-weight: 300;
-  margin-left: 12px;
-  padding: 0 11px 0 13px;
-  text-overflow: ellipsis;
-  width: 400px;
-}
-
-#pac-input:focus {
-  border-color: #4d90fe;
-}
-
-#title {
-  color: #fff;
-  background-color: #4d90fe;
-  font-size: 25px;
-  font-weight: 500;
-  padding: 6px 12px;
-}
-#target {
-  width: 345px;
-}
-
-.map-container {
-    height: 400px;
-    width: 500px;
-}
-#map {
-  height: 100%;
-}
-
-/* Google maps api css end */
-
-#alert_error,#alert_success,#alert_loading {
-  position: fixed;
-  top: 10px;
-  left: 31%;
-  width: 500px;
-  max-height: 50px;
-  font-size: 16px;
-  background-color: #ffffffc7;
-  padding: 6px 10px;
-  margin: 6px;
-  border-style: solid;
-  border-color: gray;
-  border-radius: 6px;
-  z-index: 1000;
-  text-align: center;
-  font-weight: bold;
-}
- #alert_success{
-   color: green;
-}
-#alert_error{
-  color: red;
-}
-#alert_loading {
-  color: black;
-}
-
-.container-flex{
-  width: 90%;
-  margin-left: 5%;
-  margin-right: 5%;
-  min-height: 100%;
-  justify-content: center;
-  align-items: center;
-}
-
-td, th {
-  max-width: 250px;
-}
