--- js/main.js
@@ -1,14 +1,0 @@
-
-// import * as reqHelpers from './helpers/request_helpers';
-import domHelpers from './helpers/dom_helpers.js';
-import userServices from './services/user_services.js';
-import navigationComponents from './components/navigation.js';
-
-const { displayAdminNavItems, displayHomeNavItems } = navigationComponents;
-const { login, logout, register } = userServices;
-const { getElById } = domHelpers;
-if (getElById('register_submit')) getElById('register_submit').addEventListener('click', register);
-if (getElById('login_submit')) getElById('login_submit').addEventListener('click', login);
-if (getElById('logout_link')) getElById('logout_link').addEventListener('click', logout);
-displayHomeNavItems();
-displayAdminNavItems();
