--- login.html
@@ -1,48 +1,0 @@
-<!DOCTYPE html>
-<html lang="en" dir="ltr">
-  <head>
-    <meta charset="utf-8">
-    <title>iReporter</title>
-    <link rel="stylesheet" href="./css/main.css">
-    <script type="module" src="js/main.js"></script>
-  </head>
-  <body>
-    <div id="nav" class="nav">
-  <ul id="main_nav_ul">
-    <li><a href="index.html">Home</a></li>
-    <li><a href="sign_up.html">Register</a></li>
-    <li><a href="#login">Login</a></li>
-  </ul>
-</div>
-<div class="container">
-  <div id="login" class="bg">
-    <div class="form-container">
-      <h1>Login</h1>
-      <form class="sign_up_form" >
-
-        <div>
-          <label for="Username">Username</label>
-          <input id="login_username" type="text" placeholder="Username">
-        </div>
-        <div>
-          <label for="Password">Password</label>
-          <input id="login_password" type="password" placeholder="******">
-        </div>
-        <div>
-          <button type="button" id="login_submit" class="button-default"
-          name="Login">
-            <b>Submit</b>
-          </button>
-        </div>
-      </form>
-    </div>
-  </div>
-  <div class="footer">
-    <li><a href="index.html">Home</a></li>
-    <li><a href="sign_up.html">Register</a></li>
-    <li><a href="#login">Login</a></li>
-  </div>
-</div>
-<script src="https://maps.googleapis.com/maps/api/js?key=AIzaSyDLjJJ8MEg85s3CU2eyQdZ_4H5Uaz0oud0&libraries=places"></script>
-  </body>
-</html>
