--- profile.html
@@ -1,61 +1,0 @@
-<!DOCTYPE html>
-<html lang="en" dir="ltr">
-  <head>
-    <meta charset="utf-8">
-    <title>iReporter</title>
-    <link rel="stylesheet" href="./css/main.css">
-    <script type="module" src="js/profile.js"></script>
-    <script type="module" src="js/main.js"></script>
-  </head>
-  <body>
-    <div id="nav" class="nav">
-      <ul id="main_nav_ul">
-        <li><a href="index.html">Home</a></li>
-        <li><a href="view_my_quests.html">My Red Flags/Interventions</a></li>
-        <li><a href="create_quest.html">Create an incident</a></li>
-        <li class="li-right"><a id="logout_link" >Logout</a></li>
-      </ul>
-</div>
-<div class="container">
-  <div id="profile" class="container-flex">
-      <div class="form-container">
-            <h1>Your details</h1>
-              <div class="update_user_form" >
-                <div  class="form-group">
-                  <label for="Username">Username: </label>
-                  <p id="username"></p>
-                </div>
-                <div class="form-group">
-                  <label for="First Name">First Name</label>
-                  <p id="user_fname"></p>
-                </div>
-                <div class="form-group">
-                  <label for="Last Name">Last Name</label>
-                  <p id="user_lname"></p>
-                </div>
-                <div class="form-group">
-                  <label for="Other Name">Other Name</label>
-                  <p id="user_othername"></p>
-                </div>
-                <div class="form-group">
-                  <label for="Email">Email</label>
-                  <p id="user_email"></p>
-                </div>
-                <div class="form-group">
-                  <label for="Phone">Phone</label>
-                  <p id="user_phone"></p>
-                </div>
-              </div>
-      </div>
-    </div>
-
-  </div>
-  <div class="footer">
-    <li><a href="#home">Home</a></li>
-    <li><a href="#register">Register</a></li>
-    <li><a href="#login">Login</a></li>
-  </div>
-</div>
-<script src="https://maps.googleapis.com/maps/api/js?key=AIzaSyDLjJJ8MEg85s3CU2eyQdZ_4H5Uaz0oud0&libraries=places"></script>
-  </body>
-</html>
