--- sign_up.html
@@ -1,67 +1,0 @@
-
-<!DOCTYPE html>
-<html lang="en" dir="ltr">
-  <head>
-    <meta charset="utf-8">
-    <title>iReporter</title>
-    <link rel="stylesheet" href="./css/main.css">
-    <script type="module" src="js/main.js"></script>
-  </head>
-  <body>
-    <div id="nav" class="nav">
-  <ul id="main_nav_ul">
-    <li><a href="index.html">Home</a></li>
-    <li><a href="login.html">Login</a></li>
-  </ul>
-</div>
-<div class="container">
-  <div id="register" class="bg">
-  <div class="form-container">
-    <h1>Register</h1>
-        <form class="sign_up_form">
-          <div  class="form-group">
-            <label for="Username">Username</label>
-            <input id="sign_up_username" type="text" placeholder="Username">
-          </div>
-          <div class="form-group">
-            <label for="First Name">First Name</label>
-            <input id="sign_up_fname"type="text" placeholder="First Name">
-          </div>
-          <div class="form-group">
-            <label for="Last Name">Last Name</label>
-            <input id="sign_up_lname" type="text" placeholder="Last Name">
-          </div>
-          <div class="form-group">
-            <label for="Other Name">Other Name</label>
-            <input id="sign_up_othername" type="text" placeholder="Other Name">
-          </div>
-          <div class="form-group">
-            <label for="Email">Email</label>
-            <input id="sign_up_email" type="text" placeholder="Email">
-          </div>
-          <div class="form-group">
-            <label for="Phone">Phone</label>
-            <input id="sign_up_phone" type="text" placeholder="Phone">
-          </div>
-          <div class="form-group">
-            <label for="Password">Password</label>
-            <input id="sign_up_password" type="password" placeholder="******">
-          </div>
-          <div class="form-group">
-            <button class="button-default" name="Sign Up"
-            id="register_submit" >
-            <b>Sign Up</b>
-            </button>
-          </div>
-
-      </form>
-  </div>
-</div>
-  <div class="footer">
-    <li><a href="index.html">Home</a></li>
-    <li><a href="login.html">Login</a></li>
-  </div>
-</div>
-<script src="https://maps.googleapis.com/maps/api/js?key=AIzaSyDLjJJ8MEg85s3CU2eyQdZ_4H5Uaz0oud0&libraries=places"></script>
-  </body>
-</html>
