--- view_my_quests.html
@@ -1,88 +1,0 @@
-
-
-<!DOCTYPE html>
-<html lang="en" dir="ltr">
-  <head>
-    <meta charset="utf-8">
-    <title>iReporter</title>
-    <link rel="stylesheet" href="./css/main.css">
-    <script type="module" src="js/get_all_incidents.js"></script>
-    <script type="module" src="js/main.js"></script>
-  </head>
-  <body>
-    <div id="nav" class="nav">
-      <ul id="main_nav_ul">
-        <li><a href="index.html">Home</a></li>
-        <li><a href="profile.html">My Profile</a></li>
-        <li><a href="create_quest.html">Create an incident</a></li>
-        <li class="li-right"><a id="logout_link" href="">Logout</a></li>
-      </ul>
-</div>
-<div class="container">
-  <div id="quest_table" class="bg-alt">
-  <div class="table-container">
-    <br>
-    <div>
-      <h1>Statistics For Red-flags & Interventions</h1>
-      <table>
-        <thead>
-            <tr>
-                <th>Resolved Incidents</th>
-                <th>Rejected Incidents</th>
-                <th>Under Incidents</th>
-                <th>Draft Incidents</th>
-            </tr>
-        </thead>
-        <tbody>
-            <tr class=" ">
-                <td>0</td>
-                <td>0</td>
-                <td>0</td>
-                <td>0</td>
-            </tr>
-        <tbody>
-    </table>
-    </div>
-    <h1>My Red-flags & Interventions</h1>
-    <div class="">
-      <h2><b>Create an Incident</b></h2>
-      <a href="create_quest.html" class="button-default" > Create </a>
-    </div>
-    <br>
-    <div>
-      <div class="form-group">
-        <label for="Status Filter"><b>Incident Status Filter : </b> </label>
-        <select>
-          <option value="all">All</option>
-          <option value="draft">Draft</option>
-          <option value="resolved">Resolved</option>
-          <option value="rejected">Rejected</option>
-          <option value="under investigation">Under Investigation</option>
-        </select>
-      </div>
-    </div>
-    <br>
-    <table>
-        <thead>
-            <tr>
-                <th>Incident Title</th>
-                <th>Type</th>
-                <th>Status</th>
-                <th>Actions</th>
-            </tr>
-        </thead>
-        <tbody id="incidents_tbody">
-        <tbody>
-    </table>
-  </div>
-</div>
-  <div class="footer">
-    <li><a href="index.html">Home</a></li>
-    <li><a href="profile.html">My Profile</a></li>
-    <li><a href="create_quest.html">Create an incident</a></li>
-    <li class="li-right"><a id="logout_link" href="">Logout</a></li>
-  </div>
-</div>
-<script src="https://maps.googleapis.com/maps/api/js?key=AIzaSyDLjJJ8MEg85s3CU2eyQdZ_4H5Uaz0oud0&libraries=places"></script>
-  </body>
-</html>
